requirements.txt
Flask==3.0.3
numpy==1.26.0
Pillow==9.5.0
gunicorn==23.0.0
imageio==2.31.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
mediapipe==0.10.18
libclang==18.1.1

